+++ InfinityHammer/commands/HammerRepair.cs
@@ -1,5 +1,4 @@
 using ServerDevcommands;
-using UnityEngine;
 
 namespace InfinityHammer;
 public class HammerRepairCommand
@@ -10,9 +9,8 @@ public class HammerRepairCommand
     Helper.Command("hammer_repair", "Selects the repair tool.", (args) =>
     {
       HammerHelper.Init();
-      var player = Helper.GetPlayer();
-      player.SetSelectedPiece(new Vector2Int(0, 0));
-      player.SetupPlacementGhost();
+      Hammer.SelectRepair();
+      Helper.GetPlayer().SetupPlacementGhost();
     });
   }
 }