+++ InfinityHammer/selection/BaseSelection.cs
@@ -57,6 +57,7 @@ public class BaseSelection
   }
   public virtual void Deactivate()
   {
+    Hammer.SelectRepairIfEmpty();
   }
 
 }