+++ publish/CHANGELOG.md
@@ -1,6 +1,9 @@
 - v1.49
   - Fixes multiselection not having snap points.
   - Fixes errors when multiselecting a single object.
+  - Fixes not being able to remove zoop binds.
+  - Fixes zooping creating more and more description.
+  - Fixes command rotation only working one direction.
 
 - v1.48
   - Adds support for the command `hammer_stack` to work on the hovered piece.